Move SafeDialer from cimd to httputil
The dialer that refuses to connect to a non-publicly-routable address is
not specific to fetching a client ID metadata document; the same rule has
to apply wherever a URL comes from outside this deployment. It has no
dependency on the cimd package, so move it to pkg/util/httputil.

cimd keeps its own spelling through a type alias, so its tests, which
cover the resolution and rebinding rules in detail, are unchanged.

No behaviour change.

ref DEV-3825
